fix: fix render before permission init
